pyproject.toml
[tool.poetry]
name = "test_task_gitea"
version = "0.1.0"
description = ""
authors = ["Maria Ryabova <mvryabova2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.3"
selene = "^2.0.0rc3post3"
selenium = "^4.16.0"
allure-pytest = "^2.13.2"
pytest-docker = "^2.0.1"
beautifulsoup4 = "^4.12.2"
allure-python-commons = "^2.13.2"

[tool.pytest.ini_options]
addopts = "--clean-alluredir --alluredir allure_results"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
